adding IList to xdsl.utils - add missing ignore comment
